modularize the game
Split the classes out into different modules and set up imports
accordingly. I figured out how to do this from Hard Glitch
(https://github.com/gamkedo-la/hardglitch/blob/master/index.html) via
the r/roguelikedev subreddit.
