+++ index.html
@@ -42,10 +42,6 @@
       for a bit more explaination.
     </p>
     <script src="https://cdn.jsdelivr.net/npm/mainloop.js@latest/build/mainloop.min.js"></script>
-    <script src="src/game.js"></script>
-    <script>
-      const g = new Game(80, 50);
-      MainLoop.setUpdate(d => g.update(d)).setDraw(i => g.draw(i)).start();
-    </script>
+    <script type="module" src="src/game.js"></script>
   </body>
 </html>
